fix: change car getById route
This is done because there was an error when fetching /api/car/search
and /api/car/searchAvailableCar. This is because /api/car/:id has been
implemented to fetch car by id, thus "search" and "searchAvailableCar"
is perceived to be :id. Changing /api/car/:id to /api/car/id/:id solves
this issue